Refactor index file a bit
