+++ src/api/index.php
@@ -3,207 +3,19 @@
 require_once 'lib/user.php';
 require_once 'lib/container.php';
 require_once 'lib/rqlStorage.php';
+require_once 'config.php';
+require_once 'migrations.php';
 
 use Aerys\{
 	Host, Request, Response, Router, Websocket, function root, function router, function websocket
 };
 
-define( 'DB_NAME', getenv( 'DB_NAME' ) ?: 'converser' );
-define( 'DB_HOST', getenv( 'DB_HOST' ) ?: 'localhost' );
-define( 'SMS', getenv( 'SMS' ) ?: '18037143889' );
-define( 'CALL', getenv( 'CALL' ) ?: '18882660156' );
-define( 'HOST', getenv( 'CALL_HOST' ) ?: 'http://dev.converser.space:2200/' );
-
-define( 'STRIPE_KEY', getenv( 'STRIPE_KEY' ) ?: 'sk_test_osM11tRI7n2u8cChs2J3R4kx' );
-
-define( 'METRICS_DB', DB_NAME . '_metrics' );
-
-$auth_id    = getenv( 'PLIVO_ID' ) ?: "SANTC0YTLLZTFMMZA3MM";
-$auth_token = getenv( 'PLIVO_TOKEN' ) ?: "MzA2M2UyMWViNTI5NjFmZjNjMmJiYmZlNmM5YmZh";
-
-global $conn;
-$conn = r\connect( DB_HOST );
-
-function prep() {
-	global $conn;
-	$dbs      = yield r\dbList()->run( $conn );
-	$filtered = array_filter( $dbs, function ( $db ) {
-		return $db == DB_NAME;
-	} );
-
-	if ( count( $filtered ) == 0 ) {
-		yield r\dbCreate( DB_NAME )->run( $conn );
-		$db = r\db( DB_NAME );
-		yield $db->tableCreate( 'version' )->run( $conn );
-		yield $db->table( 'version' )->insert( [
-			'id'    => 'db',
-			'value' => 0
-		] )->run( $conn );
-	}
-
-	$currentVersion = ( yield r\db( DB_NAME )->table( 'version' )->get( 'db' )->run( $conn ) )['value'];
-
-	$db   = r\db( DB_NAME );
-	$rnad = rand();
-	yield $db->table( 'version' )->insert( [ 'id' => 'rlock', 'value' => $rnad ] )->run( $conn );
-	$hasLock = ( yield $db->table( 'version' )->get( 'rlock' )->run( $conn ) )['value'] == $rnad;
-
-	var_dump( $hasLock );
-
-	if ( ! $hasLock ) {
-		return $conn;
-	}
-
-	if ( $currentVersion == null ) {
-		yield $db->table( 'version' )->insert( [
-			'id'    => 'db',
-			'value' => 0
-		] )->run( $conn );
-		$currentVersion = 0;
-	}
-
-	// todo: increment this if you add a migration
-	$expectedVersion = 13;
-	// put migrations here
-	switch ( $currentVersion + 1 ) {
-		case 1:
-			yield $db->tableCreate( 'users' )->run( $conn );
-		case 2:
-			yield $db->table( 'users' )->indexCreate( 'phone' )->run( $conn );
-		case 3:
-			yield $db->table( 'users' )->insert( [
-				'phone'   => '19102974810',
-				'admin'   => true,
-				'lives'   => 1000,
-				'score'   => 0,
-				'status'  => 'not-playing',
-				'created' => r\now()
-			] )->run( $conn );
-		case 4:
-			yield $db->tableCreate( 'calls' )->run( $conn );
-		case 5:
-			yield $db->tableCreate( 'sessions' )->run( $conn );
-		case 6:
-			yield $db->tableCreate( 'events', [ 'durability' => 'soft' ] )->run( $conn );
-		case 7:
-			yield $db->tableCreate( 'sms', [ 'durability' => 'soft' ] )->run( $conn );
-		case 8:
-			yield $db->table( 'sessions' )->indexCreate( 'phone' )->run( $conn );
-			yield $db->table( 'sessions' )->indexCreate( 'token' )->run( $conn );
-			yield $db->table( 'sessions' )->indexCreate( 'user_id' )->run( $conn );
-		case 9:
-			yield r\dbCreate( 'records' )->run( $conn );
-		case 10:
-			yield r\db( 'records' )->tableCreate( 'events' )->run( $conn );
-		case 11:
-			yield r\db( 'records' )->table( 'events' )->indexCreate( 'model_id' )->run( $conn );
-		case 12:
-			yield r\db( 'records' )->tableCreate( 'snapshots' )->run( $conn );
-		case 13:
-			yield $db->tableCreate( 'payments' )->run( $conn );
-	}
-
-	if ( $currentVersion != $expectedVersion ) {
-		yield r\db( DB_NAME )->table( 'version' )->update( [
-			'id'    => 'db',
-			'value' => $expectedVersion
-		] )->run( $conn );
-	}
-
-	yield $db->table( 'version' )->get( 'rlock' )->delete()->run( $conn );
-
-	return $conn;
-}
-
-/**
- * Call when a user first connects to the api
- *
- * @param $campaign string The Campaign (utm_)
- */
-function acquire( $campaign ) {
-}
-
-/**
- * Call when a user logins for the very first time
- *
- * @param $userid
- */
-function activate( $userid, $campaign ) {
-}
-
-/**
- * Call whenever a user returns
- *
- * @param $userid
- */
-function retented( $userid ) {
-}
-
-/**
- * Call whenever a user refers another user
- *
- * @param $from
- * @param $to
- */
-function referral( $from, $to ) {
-}
-
-/**
- * Call whenever a user generates revenue
- *
- * @param $userid
- * @param $amount
- */
-function revenue( $userid, $amount ) {
-}
-
-global $plivo;
-
-function event( $event ) {
-	$event['time'] = r\now();
-	$conn          = r\connect( DB_HOST );
-	r\db( DB_NAME )->table( 'events' )->insert( $event )->run( $conn );
-}
-
-/* --- Global server options -------------------------------------------------------------------- */
-
-const AERYS_OPTIONS = [
-	"keepAliveTimeout"   => 60,
-	"user"               => "nobody",
-	"defaultContentType" => "application/json",
-	//"deflateMinimumLength" => 0,
-];
-
-$plivo = new \Plivo\RestAPI( $auth_id, $auth_token );
-unset( $auth_id );
-unset( $auth_token );
 /* --- http://localhost:1337/ ------------------------------------------------------------------- */
-
-function PlaceCall( $from, $left, $right ) {
-	$params = [
-		'to'         => "$left<$right",
-		'from'       => $from,
-		'answer_url' => 'somewhere.com/'
-	];
-}
 
 $router = router()
 	->get( "/", function ( Request $req, Response $res ) {
 		$res->end( "<html><body><h1>Hello, world. yo...</h1></body></html>" );
 	} );
-
-global $container;
-$container            = new \Model\Container();
-$container->snapshots = r\db( 'records' )->table( 'snapshots' );
-$container->records   = r\db( 'records' )->table( 'events' );
-$container->conn      = $conn;
-$container->plivo     = $plivo;
-$container->uuid      = r\uuid();
-$container->R         = r\db( DB_NAME );
-$container->charge    = 'Stripe\Charge';
-$container->storage   = new \Model\RqlStorage( $container );
-
-Stripe\Stripe::setApiKey( STRIPE_KEY );
 
 $websocket = websocket( new class implements Aerys\Websocket {
 	/**
@@ -233,81 +45,6 @@ $websocket = websocket( new class implements Aerys\Websocket {
 	 */
 	private function cleanPhone( string $phone ): string {
 		return preg_replace( '/\D+/', '', $phone );
-	}
-
-	private function getRank( $phone ) {
-		global $conn;
-		$phone = $this->cleanPhone( $phone );
-		$rank  = r\db( DB_NAME )
-			->table( 'events' )
-			->filter( [
-				'type' => 'increase_score',
-			] )
-			->group( 'player' )
-			->map( function ( $row ) {
-				return $row( 'amount' );
-			} )
-			->ungroup()
-			->map( function ( $res ) {
-				return r\expr( [
-					'player' => $res( 'group' ),
-					'score'  => $res( 'reduction' )->sum()
-				] );
-			} )
-			->orderBy( r\desc( 'score' ) )
-			->offsetsOf( function ( $row ) use ( $phone ) {
-				return $row( 'player' )->eq( $phone );
-			} )->run( $conn )->toArray();
-
-		if ( empty( $rank ) ) {
-			return 'not ranked';
-		}
-
-		return $rank[0];
-	}
-
-	private function getLeaderboard() {
-		global $conn;
-		$lastDay = r\db( DB_NAME )
-			->table( 'events' )
-			->filter( [ 'type' => 'increase_score' ] )
-			->group( 'player' )
-			->map( function ( $row ) {
-				return $row( 'amount' );
-			} )
-			->ungroup()
-			->map( function ( $res ) {
-				return r\expr( [
-					'player_prefix' => $res( 'group' )->slice( 1, 4 ),
-					'player_suffix' => $res( 'group' )->slice( - 4 ),
-					'score'         => $res( 'reduction' )->sum()
-				] );
-			} )
-			->orderBy( r\desc( 'score' ) )
-			->limit( 10 )->run( $conn )->toArray();
-
-		$allTime = r\db( DB_NAME )
-			->table( 'events' )
-			->filter( [ 'type' => 'increase_score' ] )
-			->group( 'player' )
-			->map( function ( $row ) {
-				return $row( 'amount' );
-			} )
-			->ungroup()
-			->map( function ( $res ) {
-				return r\expr( [
-					'player_prefix' => $res( 'group' )->slice( 1, 4 ),
-					'player_suffix' => $res( 'group' )->slice( - 4 ),
-					'score'         => $res( 'reduction' )->sum()
-				] );
-			} )
-			->orderBy( r\desc( 'score' ) )
-			->limit( 10 )->run( $conn )->toArray();
-
-		return [
-			'lastDay' => $lastDay,
-			'allTime' => $allTime
-		];
 	}
 
 	/**
@@ -416,7 +153,7 @@ $websocket = websocket( new class implements Aerys\Websocket {
 					unset( $user );
 					break;
 				case 'connect':
-					acquire( $request['campaign'] );
+					//acquire( $request['campaign'] );
 					break;
 			}
 		}
